refactor: improve medical attributes structure and resolve issues
- Fix enum validation in User.cs and BloodType.cs (correct range from 1 to Length)
- Update tests to check exception types only, removing message assertions for better refactoring resistance
- Resolve Swagger schema ID conflicts by renaming DTOs in self endpoints:
  - SurgeryDto -> SurgerySummaryDto
  - MedicationDto -> MedicationSummaryDto
  - AllergyDto -> AllergySummaryDto
  - ChronicDiseaseDto -> ChronicDiseaseSummaryDto
- Add OpenAPI groups for medical attribute endpoints:
  - AllergiesGroup, ChronicDiseasesGroup, MedicationsGroup, SurgeriesGroup
- All 131 tests now passing

+++ src/MedicalCenter.Infrastructure/Services/TokenProvider.cs
@@ -28,7 +28,7 @@ public class TokenProvider(IOptions<JwtSettings> jwtSettings) : ITokenProvider
             new(JwtRegisteredClaimNames.Name, user.FullName),
             new(ClaimTypes.Role, user.Role.ToString()),
             new("userId", user.Id.ToString()),
-            new("userRole", ((int)user.Role).ToString())
+            new("userRole", user.Role.ToString())
         };
 
         // Add role-specific claims
